🏗️ Architect: Modularized GuideScope and consolidated Barlow
Refactored the 1,143-line `apts/opticalequipment/guide_scope.py` into a modular package structure in `apts/opticalequipment/guide_scope/`.
- Core logic isolated in `base.py`.
- 65 factory methods and database entries distributed across 29 vendor-specific modules in `vendors/`.
- Dynamic aggregation in `__init__.py` maintains the public API.

Consolidated the `Barlow` module by removing the redundant 1,674-line `apts/opticalequipment/barlow.py` file, as its functionality is fully covered by the `apts/opticalequipment/barlow/` package.

Verified all changes with a comprehensive 660-test unit suite.
