+++ apts/opticalequipment/guide_scope/vendors/takahashi.py
@@ -1,0 +1,23 @@
+from ..base import GuideScope
+
+
+class TakahashiGuideScope(GuideScope):
+    _DATABASE = {
+        "Takahashi_GT_40_Guide_Scope": {
+            "brand": "Takahashi",
+            "name": "GT-40 Guide Scope",
+            "type": "type_guide_scope",
+            "optical_length": 0,
+            "mass": 250,
+            "tside_thread": "",
+            "tside_gender": "",
+            "cside_thread": "M42",
+            "cside_gender": "Female",
+            "reversible": False,
+            "bf_role": "",
+        },
+    }
+
+    @classmethod
+    def Takahashi_GT_40_Guide_Scope(cls):
+        return cls.from_database(cls._DATABASE["Takahashi_GT_40_Guide_Scope"])
