+++ apts/opticalequipment/guide_scope/vendors/tecnosky.py
@@ -1,0 +1,40 @@
+from ..base import GuideScope
+
+
+class TecnoskyGuideScope(GuideScope):
+    _DATABASE = {
+        "Tecnosky_30mm_Guide_Scope": {
+            "brand": "Tecnosky",
+            "name": "30mm Guide Scope",
+            "type": "type_guide_scope",
+            "optical_length": 0,
+            "mass": 110,
+            "tside_thread": "",
+            "tside_gender": "",
+            "cside_thread": "CS",
+            "cside_gender": "Female",
+            "reversible": False,
+            "bf_role": "",
+        },
+        "Tecnosky_60mm_Guide_Scope": {
+            "brand": "Tecnosky",
+            "name": "60mm Guide Scope",
+            "type": "type_guide_scope",
+            "optical_length": 0,
+            "mass": 310,
+            "tside_thread": "",
+            "tside_gender": "",
+            "cside_thread": "M42",
+            "cside_gender": "Female",
+            "reversible": False,
+            "bf_role": "",
+        },
+    }
+
+    @classmethod
+    def Tecnosky_30mm_Guide_Scope(cls):
+        return cls.from_database(cls._DATABASE["Tecnosky_30mm_Guide_Scope"])
+
+    @classmethod
+    def Tecnosky_60mm_Guide_Scope(cls):
+        return cls.from_database(cls._DATABASE["Tecnosky_60mm_Guide_Scope"])
